refactor: errorHandler do app.ts
